requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
web3>=6.11.1
requests>=2.31.0
aiohttp>=3.9.1
python-dateutil>=2.8.2